Fix database configuration file
It does not make sense to have the FETCH_MODE specified there.
This config file will not change the system behavior, only
specify connection settings. The db directive was not being
used.

Also, the redundant include directory was removed, since the config file
that was being used was in the conf directory.